Refactors and moves send method to Pocketeer main module
This allows a single module to be used to add, modify or fetch items from Pocket. The helper structs `Item` and `Get` can still be used in conjunction with the API functions.
